requirements.txt

google-generativeai
python-dotenv
langchain
sentence-transformers
scikit-learn
matplotlib
numpy
